% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm-package.R
\docType{package}
\name{rusdm-package}
\alias{rusdm-package}
\alias{rusdm}
\title{rusdm-package: Access data from the US Drought Monitor API}
\description{
rusdm is a wrapper for the United States Department of Agriculture's
National Agricultural Statistical Service (NASS) 'Quick Stats' API to enable
getting NASS 'Quick Stats' data directly from \R.  Based on the httr API
package guide.
}
\details{
The functions in this package facilitate getting data from NASS 'Quick Stats'.
It handles the API key checking and storage, authorization, and fetching of
data.
}
\references{
\url{https://quickstats.nass.usda.gov}
}
\seealso{
\url{https://quickstats.nass.usda.gov/api}
}
\author{
Nicholas Potter
}
\keyword{internal}
\keyword{package}
\keyword{rusdm-package}
