% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{usdm}
\alias{usdm}
\title{Get USDM data}
\usage{
usdm(
  area,
  stat,
  aoi = NULL,
  startdate,
  enddate,
  dx = NULL,
  DxLevelThresholdFrom = NULL,
  DxLevelThresholdTo = NULL,
  minimumweeks = NULL,
  hucLevel = NULL,
  statisticsType = c("traditional", "categorical"),
  progress_bar = TRUE,
  format = "CSV"
)
}
\arguments{
\item{area}{the area of data to query. See \link{usdm_areas}.}

\item{stat}{the statistic to query. See \link{usdm_stats}.}

\item{aoi}{Area of interest to query. This depends on the \code{area}, and is
generally the ID number of the area of interest. If \code{area = National},
\code{conus} queries the contiguous U.S., \code{total} queries the total U.S., and
\code{us} queries both. States are the two-digit FIPS code. Counties are the
five-digit FIPS code, or the two-letter state abbreviation in order to get
all counties in a state. See the \href{https://droughtmonitor.unl.edu/DmData/DataDownload/WebServiceInfo.aspx}{USDM API}
for details.}

\item{startdate}{the start date for the query, formatted as \code{M/D/YYYY}, e.g.
"1/1/2012" or "12/31/2024".}

\item{enddate}{the end date for the query, formatted as \code{M/D/YYYY}, e.g.
"1/1/2012" or "12/31/2024".}

\item{dx}{minimum drought level from 0 to 4, corresponding to drought categories D0-D4.}

\item{DxLevelThresholdFrom}{minimum drought threshold to be considered "drought", from 0-4.}

\item{DxLevelThresholdTo}{maximum drought threshold to be considered "drought", from 0-4.}

\item{minimumweeks}{minimum number of weeks to be considered a drought. Only
applicable for \code{\link[=usdm_weeks]{usdm_weeks()}} or \code{area = ConsecutiveNonConsecutiveStatistics}.}

\item{hucLevel}{The watershed (HUC) resolution. Either 2, 4, 6, or 8. Only applies
if \code{area = HUCStatistics}.}

\item{statisticsType}{either "traditional" or "categorical".}

\item{progress_bar}{whether to display a download progress bar.}

\item{format}{the format of data returned, either "CSV", "XML", or "JSON".}
}
\value{
a data.frame or unparsed object.
}
\description{
This is the generic query function to get data using the US Drought Monitor's
api. There are five statistics for which to query drought classifications,
each of which has an associated function that wraps this function. For
the area in each drought classification, use \code{\link[=usdm_area]{usdm_area()}}. For the percent
of area, use \code{\link[=usdm_areapercent]{usdm_areapercent()}}. For population and percent of population
use \code{\link[=usdm_population]{usdm_population()}} and \code{\link[=usdm_populationpercent]{usdm_populationpercent()}}. Use \code{\link[=usdm_dcsi]{usdm_dcsi()}} for
the DSCI.
}
\seealso{
wrappers \code{\link[=usdm_area]{usdm_area()}}, \code{\link[=usdm_areapercent]{usdm_areapercent()}}, \code{\link[=usdm_population]{usdm_population()}},
\code{\link[=usdm_populationpercent]{usdm_populationpercent()}}, \code{\link[=usdm_dcsi]{usdm_dcsi()}} for querying specific statistics.
}
