% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{usdm_check}
\alias{usdm_check}
\title{Check a query response.}
\usage{
usdm_check(response)
}
\arguments{
\item{response}{a \code{\link[httr:GET]{httr::GET()}} request result returned from the API.}
}
\value{
nothing if check is passed, or an informative error if not passed.
}
\description{
Check a query response.
}
