% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{usdm_dsci}
\alias{usdm_dsci}
\title{Get the DSCI in different drought classifications.}
\usage{
usdm_dsci(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=usdm]{usdm()}}.}
}
\value{
a data.frame or other response from the USDM API.
}
\description{
Get the DSCI in different drought classifications.
}
