% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{usdm_GET}
\alias{usdm_GET}
\title{Send a query to the US Drought Monitor API.}
\usage{
usdm_GET(
  params,
  api_path,
  progress_bar = TRUE,
  format = c("CSV", "JSON", "XML")
)
}
\arguments{
\item{params}{a named list of parameters composed of \code{aoi}, \code{startdate},
\code{enddate}, \code{statisticsType}, and \code{hucLevel}, which are set by the
corresponding parameters in \code{\link[=usdm]{usdm()}}.}

\item{api_path}{(char) api path set by <\code{area}/\code{stat}> in \code{\link[=usdm]{usdm()}}.}

\item{progress_bar}{whether to display a download progress bar.}

\item{format}{the format of data returned, either "CSV", "XML", or "JSON".}
}
\value{
an httr response object.
}
\description{
Send a query to the US Drought Monitor API.
}
