% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{usdm_parse}
\alias{usdm_parse}
\title{Parse a query from the US Drought Monitor API.}
\usage{
usdm_parse(req, as_numeric = TRUE, as = c("data.frame", "list", "text"), ...)
}
\arguments{
\item{req}{a request made by \code{\link[httr:GET]{httr::GET()}}.}

\item{as_numeric}{convert dates and numerics from character to date and
numeric format.}

\item{as}{whether to parse the request as a data.frame, list, or plain text.}

\item{...}{additional parameters for \code{\link[=read.csv]{read.csv()}} or
\code{\link[jsonlite:read_json]{jsonlite::parse_json()}}.}
}
\value{
a data.frame, list, or plain text object.
}
\description{
Parse a query from the US Drought Monitor API.
}
