% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{usdm_population}
\alias{usdm_population}
\title{Get the population of area in different drought classifications.}
\usage{
usdm_population(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=usdm]{usdm()}}.}
}
\value{
a data.frame or other response from the USDM API.
}
\description{
Get the population of area in different drought classifications.
}
