% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{usdm_populationpercent}
\alias{usdm_populationpercent}
\title{Get the percent of population in different drought classifications.}
\usage{
usdm_populationpercent(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=usdm]{usdm()}}.}
}
\value{
a data.frame or other response from the USDM API.
}
\description{
Get the percent of population in different drought classifications.
}
