% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{usdm_weeks}
\alias{usdm_weeks}
\title{Get consecutive or non-consecutive weeks in a drought category.}
\usage{
usdm_weeks(..., consecutive = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{\link[=usdm]{usdm()}}.}

\item{consecutive}{whether to query consecutive or non-consecutive weeks.}
}
\value{
a data.frame or other response from the USDM API.
}
\description{
Get consecutive or non-consecutive weeks in a drought category.
}
